requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
numpy>=1.26.0
pandas>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0.1
transformers>=4.30.0
